Change vmMain to use intptr_t

+++ src/qcommon/qcommon.h
@@ -1033,7 +1033,7 @@ void Sys_LeaveCriticalSection( void *ptr );
 
 // general development dll loading for virtual machine testing
 
-void    * QDECL Sys_LoadDll( const char *name, intptr_t( QDECL * *entryPoint ) ( int, ... ),
+void    * QDECL Sys_LoadDll( const char *name, intptr_t( QDECL * *entryPoint ) ( intptr_t, ... ),
 							 intptr_t ( QDECL * systemcalls )( intptr_t, ... ) );
 void    Sys_UnloadDll( void *dllHandle );
