+++ src/qcommon/vm_local.h
@@ -142,7 +142,7 @@ struct vm_s {
 
 	// for dynamic linked modules
 	void        *dllHandle;
-	intptr_t ( QDECL *entryPoint )( int callNum, ... );
+	intptr_t ( QDECL *entryPoint )( intptr_t callNum, ... );
 
 	// for interpreted modules
 	qboolean currentlyInterpreting;
